Generates simulated data with standardized parameters.

This function takes a lavaan model with standardized parameters and simulates latent scores, errors, disturbances, and observed scores.

sim_standardized(m, n = 1000, observed = TRUE, latent = TRUE,
  errors = TRUE, factor_scores = FALSE, composites = FALSE,
  matrices = FALSE)

Arguments

m
Structural model represented by lavaan syntax

n
Number of simulated cases

observed
Include observed variables

latent
Include latent variables

errors
Include observed error and latent disturbances variables

factor_scores
Include factor score variables

composites
Include composite variables

matrices
Include matrices as attribute of tibble

Value

tibble with standardized data

Details

This function supports the `~` operator for regressions, the `~~` for covariances (but not variances), and the `=~` latent variable loadings. It does not support intercepts (e.g,. `y ~ 1`), thresholds, scaling factors, formative factors, or equality constraints.

Examples

library(simstandard)
# Lavaan model
m = "Latent_1 =~ 0.8 * Ob_1 + 0.8 * Ob_2"

# simulate 10 cases
sim_standardized(m, n = 10)

#> # A tibble: 10 x 5
#>      Ob_1   Ob_2 Latent_1   e_Ob_1  e_Ob_2
#>     <dbl>  <dbl>    <dbl>    <dbl>   <dbl>
#>  1 -2.79  -1.80   -2.44   -0.840    0.153 
#>  2  0.915  1.29    1.15   -0.00334  0.373 
#>  3 -1.29  -0.344  -0.244  -1.09    -0.148 
#>  4  0.334  0.171   0.629  -0.170   -0.332 
#>  5  1.65  -0.569   0.512   1.24    -0.979 
#>  6 -1.16  -0.355  -0.0526 -1.12    -0.313 
#>  7  0.700 -0.174   0.468   0.326   -0.548 
#>  8  0.808 -0.193   0.738   0.218   -0.783 
#>  9  0.384 -0.807  -0.936   1.13    -0.0585
#> 10 -1.22  -1.71   -1.51   -0.00957 -0.496