Function that takes a lavaan model with standardized parameters returns a list with model caracteristics

Function that takes a lavaan model with standardized parameters returns a list with model caracteristics

sim_standardized_matrices(m, max_iterations = 100)

Arguments

m
Structural model represented by lavaan Syntax

max_iterations
Maximum number of iterations before the algorithm fails

Value

list of path and covariance coefficients

Examples

# lavaan model
m = "Latent_1 =~ 0.8 * Ob_1 + 0.8 * Ob_2"

sim_standardized_matrices(m)

#> $RAM_matrices
#> $RAM_matrices$A
#>          Ob_1 Ob_2 Latent_1
#> Ob_1        0    0      0.8
#> Ob_2        0    0      0.8
#> Latent_1    0    0      0.0
#> 
#> $RAM_matrices$S
#>          Ob_1 Ob_2 Latent_1
#> Ob_1     0.36 0.00        0
#> Ob_2     0.00 0.36        0
#> Latent_1 0.00 0.00        1
#> 
#> $RAM_matrices$filter_matrix
#>          Ob_1 Ob_2 Latent_1
#> Ob_1        1    0        0
#> Ob_2        0    1        0
#> Latent_1    0    0        0
#> 
#> $RAM_matrices$iA
#>          Ob_1 Ob_2 Latent_1
#> Ob_1        1    0      0.8
#> Ob_2        0    1      0.8
#> Latent_1    0    0      1.0
#> 
#> 
#> $Correlations
#> $Correlations$R
#>          Ob_1 Ob_2 Latent_1
#> Ob_1     1.00 0.64      0.8
#> Ob_2     0.64 1.00      0.8
#> Latent_1 0.80 0.80      1.0
#> 
#> $Correlations$R_all
#>                         Ob_1      Ob_2  Latent_1     e_Ob_1     e_Ob_2
#> Ob_1               1.0000000 0.6400000 0.8000000  0.6000000  0.0000000
#> Ob_2               0.6400000 1.0000000 0.8000000  0.0000000  0.6000000
#> Latent_1           0.8000000 0.8000000 1.0000000  0.0000000  0.0000000
#> e_Ob_1             0.6000000 0.0000000 0.0000000  1.0000000  0.0000000
#> e_Ob_2             0.0000000 0.6000000 0.0000000  0.0000000  1.0000000
#> Latent_1_FS        0.9055385 0.9055385 0.8834522  0.3312946  0.3312946
#> e_Ob_1_FS          0.7683749 0.0000000 0.3748170  0.7808688 -0.4997560
#> e_Ob_2_FS          0.0000000 0.7683749 0.3748170 -0.4997560  0.7808688
#> Latent_1_Composite 0.9055385 0.9055385 0.8834522  0.3312946  0.3312946
#>                    Latent_1_FS  e_Ob_1_FS  e_Ob_2_FS Latent_1_Composite
#> Ob_1                 0.9055385  0.7683749  0.0000000          0.9055385
#> Ob_2                 0.9055385  0.0000000  0.7683749          0.9055385
#> Latent_1             0.8834522  0.3748170  0.3748170          0.8834522
#> e_Ob_1               0.3312946  0.7808688 -0.4997560          0.3312946
#> e_Ob_2               0.3312946 -0.4997560  0.7808688          0.3312946
#> Latent_1_FS          1.0000000  0.4242641  0.4242641          1.0000000
#> e_Ob_1_FS            0.4242641  1.0000000 -0.6400000          0.4242641
#> e_Ob_2_FS            0.4242641 -0.6400000  1.0000000          0.4242641
#> Latent_1_Composite   1.0000000  0.4242641  0.4242641          1.0000000
#> 
#> 
#> $Coefficients
#> $Coefficients$factor_score
#>      Latent_1_FS  e_Ob_1_FS  e_Ob_2_FS
#> Ob_1   0.4878049  1.0162602 -0.6504065
#> Ob_2   0.4878049 -0.6504065  1.0162602
#> 
#> $Coefficients$composite_score
#>      Latent_1_Composite
#> Ob_1          0.5521576
#> Ob_2          0.5521576
#> 
#> 
#> $v_names
#> $v_names$v_observed
#> [1] "Ob_1" "Ob_2"
#> 
#> $v_names$v_latent
#> [1] "Latent_1"
#> 
#> $v_names$v_latent_exogenous
#> [1] "Latent_1"
#> 
#> $v_names$v_latent_endogenous
#> character(0)
#> 
#> $v_names$v_observed_exogenous
#> character(0)
#> 
#> $v_names$v_observed_endogenous
#> [1] "Ob_1" "Ob_2"
#> 
#> $v_names$v_observed_indicator
#> [1] "Ob_1" "Ob_2"
#> 
#> $v_names$v_disturbance
#> character(0)
#> 
#> $v_names$v_error
#> [1] "e_Ob_1" "e_Ob_2"
#> 
#> $v_names$v_residual
#> [1] "e_Ob_1" "e_Ob_2"
#> 
#> $v_names$v_factor_score
#> [1] "Latent_1_FS" "e_Ob_1_FS"   "e_Ob_2_FS"  
#> 
#> $v_names$v_composite_score
#> [1] "Latent_1_Composite"
#> 
#> 
#> $iterations
#> [1] 2
#>